implement map and write feature
this is slightly more general than the existing "write" functionality, since
it can apply a function to a new dataset before writing.

This can be used to write a simple apache beam transformation.

add .partition method to return a list

Mathematically, a partition of a set A is a list disjoint sets
(potentially infinite) that cover A. The current code encodes this list
like structure using "dims", "rank", "ranks", but given how tightly coupled
this concepts are (e.g. rank < ranks) it makes sense to collect them in
one data structure, in this case a list of regions.