Fix invalid rank error (#12)
The previous implementation of the logic to catch whether a `rank` passed to `partition.indexers` was valid was incorrect.  The minus one is on the wrong side of the inequality; it should be either `rank > ranks - 1`, or, as I've implemented things now, `rank >= ranks`.  This PR adds a test and fixes this issue.  

In practice this was not a big deal, because it just meant that sometimes we would not trigger the error when we should have.